VBR support, moved minimal client parts to be shared with standard client
